Add more sophisticated hysteresis when switching to/from encoder
With borderline observer tuning and sensorless handover speed, I
observed oscillations between the observer and the encoder because the
difference in position would change the calculated velocity enough to
switch back. This oscillation resulted in failing to produce enough
torque to increase the speed enough to fully switch to sensorless, but
using either the encoder or the observer would work well enough to keep
going. To fix this, track speeds separately for the observer and the
encoder, and use the appropriate speed when deciding whether to switch.
Also add a minimum time between switching to further increase the
chances of success.

Note that once I got everything tuned well, this is no longer necessary.
However, I think making things more robust is still helpful.