Swap 60_75 sampling order back
This reverts the relevant part of 58b5c85a6. The commit has nothing
useful about why this change was made, but similar changes to HW_60
happened in the previous commit which mentions HFI. There don't seem to
be any other changes in other parts of the code corresponding to this,
so we'd probably be fine either way, but we're going to stick with the
older version that's closer to our firmware anyways.